winepulse: updated pulse driver to v9
V9:
- Rework IAudioClient::Reset, flushing is only needed for render stream.
- Add stream validity checks to IAudioClock as well
- Fix copy pasted bug from winealsa in clone_format (As reported by KittyCat)
- Validate wBitsPerSample for float format (kcat)
- Make the pulse mapping const (kcat)
- Change mix format to float (kcat)
- Fix volume comments referring to alsa instead of Pulseaudio
- Split off pulseaudio spec to waveformat to a separate function
- Disabled audioclock interpolation for now, until I determine how to do it properly.
- Call SetEvent in render callback as last function, since it gives up time slot (as reported by Joerg)
- Reworked capture logic to always capture, since that seems to be the case on windows,
  when stopped we simply drop whole packets, but keep firing the event.
- Rework the logic behind GetMixFormat, and make capture and rendering formats and periods separate again.
  This is needed for example if you have 5.1 playback 44100 with 32kHz 2 channel capture (USB microphone).
- Complete the pa_channel_position -> dwChannelMask mapping
- Use MulDiv instead of pa_bytes_to_usec or pa_usec_to_bytes in a few places.
- Rendering buffers no longer have to be a multiple of period size.

There are 2 places that still have those calls, but in worst case it will make a 50 ns difference
for the reported device periods, while the one in AudioClock::GetPosition is disabled until it's
determined how to handle interpolation of IAudioClock::GetPosition, if that's still going to happen.

I believe with this I addressed all issues and comments raised in V8,
and looking forward to any constructive feedback and test results of V9.
In particular, I'm interested in XAudio2 results with windows version
set to vista or higher, so that it uses mmdevapi directly

winmm and dsound should be able to work unmodified, but you may want
to lower ds_snd_queue_max to 3 or 4 to decrease latency on dsound games.
It seems dsound doesn't respect mmdevapi in that regard or 2 periods
might have been possible. It would have been nice if it did. :(